getHomeserver js test
