+++ pubky/src/wasm/api/auth.rs
@@ -107,11 +107,11 @@ impl Client {
     pub async fn get_homeserver(
         &self,
         public_key: &PublicKey,
-    ) -> Result<JsValue, JsValue> {
+    ) -> Result<String, String> {
         let val = self.0.get_homeserver(public_key.as_inner()).await;
         match val {
-            Some(homeserver) => Ok(JsValue::from_str(&homeserver)),
-            None => Err(JsValue::from_str("No homeserver found")),
+            Some(homeserver) => Ok(homeserver),
+            None => Err(String::from("No homeserver found")),
         }
     }
 }
