--- pubky/src/wasm/api/mod.rs
@@ -1,4 +1,0 @@
-mod auth;
-mod public;
-mod http;
-mod recovery_file;
