--- pubky/src/wasm/wrappers/mod.rs
@@ -1,2 +1,0 @@
-mod keys;
-mod session;
